requirements.txt
fastapi
uvicorn
sqlalchemy
python-multipart
jinja2
pandas
requests
twilio
python-dotenv
openai
httpx
pyzipper
websockets
pydantic-settings
psycopg2-binary
